+++ storage/common.go
